+++ qiskit/synthesis/__init__.py
@@ -89,6 +89,10 @@ Basis Change Synthesis
 
 .. autofunction:: synth_qft_line
 
+Unitary Synthesis
+=================
+
+.. autofunction:: qs_decomposition
 """
 
 from .evolution import (
@@ -126,3 +130,4 @@ from .cnotdihedral import (
 from .stabilizer import synth_stabilizer_layers, synth_stabilizer_depth_lnn
 from .discrete_basis import SolovayKitaevDecomposition, generate_basic_approximations
 from .qft import synth_qft_line
+from .unitary.qsd import qs_decomposition
