+++ qiskit/synthesis/unitary/qsd.py
